chore: refactor container and point of sale seeders
