+++ test/seed/catalogue/index.ts
@@ -16,6 +16,8 @@
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
+export { default as ContainerSeeder } from './container';
+export { default as PointOfSaleSeeder } from './point-of-sale';
 export { default as ProductSeeder } from './product';
 export { default as ProductCategorySeeder } from './product-category';
 export { default as VatGroupSeeder } from './vat-group';
