+++ test/unit/controller/authentication-secure-controller.ts
@@ -31,16 +31,12 @@ import TokenMiddleware from '../../../src/middleware/token-middleware';
 import { truncateAllTables } from '../../setup';
 import { finishTestDB } from '../../helpers/test-helpers';
 import PointOfSale from '../../../src/entity/point-of-sale/point-of-sale';
-import {
-  seedContainers,
-  seedPointsOfSale,
-} from '../../seed-legacy';
 import MemberAuthenticator from '../../../src/entity/authenticator/member-authenticator';
 import AuthenticationResponse from '../../../src/controller/response/authentication-response';
 import DefaultRoles from '../../../src/rbac/default-roles';
 import settingDefaults from '../../../src/server-settings/setting-defaults';
 import ServerSettingsStore from '../../../src/server-settings/server-settings-store';
-import { ProductSeeder, RbacSeeder, UserSeeder } from '../../seed';
+import { PointOfSaleSeeder, RbacSeeder, UserSeeder } from '../../seed';
 
 describe('AuthenticationSecureController', () => {
   let ctx: {
@@ -73,9 +69,7 @@ describe('AuthenticationSecureController', () => {
       users.filter((u) => u.type === UserType.ORGAN),
     );
 
-    const { productRevisions } = await new ProductSeeder().seedProducts(users);
-    const { containerRevisions } = await seedContainers(users, productRevisions);
-    const { pointsOfSale, pointOfSaleUsers } = await seedPointsOfSale(users, containerRevisions);
+    const { pointsOfSale, pointOfSaleUsers } = await new PointOfSaleSeeder().seedPointsOfSale(users);
 
     await DefaultRoles.synchronize();
     const roleManager = new RoleManager();
