+++ test/unit/controller/point-of-sale-controller.ts
@@ -21,10 +21,6 @@ import { json } from 'body-parser';
 import chai, { expect, request } from 'chai';
 import PointOfSaleController from '../../../src/controller/point-of-sale-controller';
 import User, { UserType } from '../../../src/entity/user/user';
-import {
-  seedContainers,
-  seedPointsOfSale,
-} from '../../seed-legacy';
 import TokenMiddleware from '../../../src/middleware/token-middleware';
 import PointOfSale from '../../../src/entity/point-of-sale/point-of-sale';
 import {
@@ -60,7 +56,7 @@ import { SwaggerSpecification } from 'swagger-model-validator';
 import { SeededRole } from '../../seed/rbac';
 import PointOfSaleService from '../../../src/service/point-of-sale-service';
 import MemberAuthenticator from '../../../src/entity/authenticator/member-authenticator';
-import { ProductSeeder, RbacSeeder, UserSeeder } from '../../seed';
+import { ContainerSeeder, PointOfSaleSeeder, RbacSeeder, UserSeeder } from '../../seed';
 
 chai.use(deepEqualInAnyOrder);
 
@@ -129,9 +125,8 @@ describe('PointOfSaleController', async () => {
     const feut2 = users.filter((u) => u.type === UserType.MEMBER)[1];
     const bestuur1 = users.filter((u) => u.type === UserType.MEMBER)[2];
 
-    const { productRevisions } = await new ProductSeeder().seedProducts([adminUser, organUser]);
-    const { containers, containerRevisions } = await seedContainers([adminUser, organUser], productRevisions);
-    const { pointsOfSale, pointOfSaleRevisions } = await seedPointsOfSale([adminUser, organUser], containerRevisions);
+    const { containers, containerRevisions } = await new ContainerSeeder().seedContainers([adminUser, organUser]);
+    const { pointsOfSale, pointOfSaleRevisions } = await new PointOfSaleSeeder().seedPointsOfSale([adminUser, organUser], containerRevisions);
 
     const all = { all: new Set<string>(['*']) };
     const own = { own: new Set<string>(['*']) };
